backend/requirements.txt
google-genai
python-dotenv
snowflake-connector-python
fastapi 
uvicorn